% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation.R
\name{build_slum_here}
\alias{build_slum_here}
\title{Adds a here sentinel file to a slumdown blog}
\usage{
build_slum_here(dir, quietly = FALSE)
}
\arguments{
\item{dir}{Where is the blog directory}

\item{quietly}{Suppress the output message (default = FALSE)}
}
\description{
Adds a here sentinel file to a slumdown blog
}
\details{
Adds the .here file to specify project root
}
