% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation.R
\name{build_slum_locally}
\alias{build_slum_locally}
\title{Build slumdown blog from local source}
\usage{
build_slum_locally(dir)
}
\arguments{
\item{dir}{string specifying the blog location}
}
\description{
Build slumdown blog from local source
}
\details{
Creates the directory structure for a new slumdown blog using
the local copy of the hugo-slum.zip file. It does not rely on internet
resources.
}
