% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation.R
\name{build_slum_nojekyll}
\alias{build_slum_nojekyll}
\title{Adds a nojekyll file to a slumdown blog}
\usage{
build_slum_nojekyll(dir, quietly = FALSE)
}
\arguments{
\item{dir}{Where is the blog directory}

\item{quietly}{Suppress the output message (default = FALSE)}
}
\description{
Adds a nojekyll file to a slumdown blog
}
\details{
Adds the .nojekyll file for github pages deployment
}
