% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation.R
\name{build_slum_project}
\alias{build_slum_project}
\title{Adds a project file to slumdown blog}
\usage{
build_slum_project(dir)
}
\arguments{
\item{dir}{Where is the blog directory}
}
\description{
Adds a project file to slumdown blog
}
\details{
Adds an RStudio project file or .here file to specify project root
}
