% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation.R
\name{build_slum_remotely}
\alias{build_slum_remotely}
\title{Build slumdown blog from remote source}
\usage{
build_slum_remotely(dir)
}
\arguments{
\item{dir}{string specifying the blog location}
}
\description{
Build slumdown blog from remote source
}
