% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caladown-package.R
\docType{package}
\name{caladown-package}
\alias{caladown}
\alias{caladown-package}
\title{caladown: A Hugo Theme for R users}
\description{
Provides a simple theme for building websites with blogdown or 
    hugodown. The caladown package contains wrapper functions to install the 
    calade theme for Hugo, ensures that the user has the necessary R packages
    required to build the site.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/djnavarro/caladown}
  \item Report bugs at \url{https://github.com/djnavarro/caladown/issues}
}

}
\author{
\strong{Maintainer}: Danielle Navarro \email{d.navarro@unsw.edu.au} (\href{https://orcid.org/0000-0001-7648-6578}{ORCID})

}
\keyword{internal}
