% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_blogdown_calade.R
\name{create_blogdown_calade}
\alias{create_blogdown_calade}
\title{Create a calade site using blogdown}
\usage{
create_blogdown_calade(path = ".", open = rlang::is_interactive(), ...)
}
\arguments{
\item{path}{Path to create site}

\item{open}{Open new site after creation?}

\item{...}{Arguments to pass to blogdown::new_site}
}
\description{
Create a calade site using blogdown
}
