% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_blogdown.R
\name{create_blogdown_slum}
\alias{create_blogdown_slum}
\title{Create a blogdown slum site}
\usage{
create_blogdown_slum(path = ".", open = rlang::is_interactive(), ...)
}
\arguments{
\item{path}{Path to create site}

\item{open}{Open new site after creation?}

\item{...}{Arguments to pass to blogdown::new_site}
}
\description{
Create site with blogdown
}
