% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_hugodown.R
\name{create_hugodown_calade}
\alias{create_hugodown_calade}
\title{Create a calade site using hugodown}
\usage{
create_hugodown_calade(
  path = ".",
  open = rlang::is_interactive(),
  rstudio = rstudioapi::isAvailable()
)
}
\arguments{
\item{path}{Path to create site}

\item{open}{Open new site after creation?}

\item{rstudio}{Create RStudio project?}
}
\description{
Create a calade site using hugodown
}
