% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hugodown_diziet.R
\name{create_hugodown_diziet}
\alias{create_hugodown_diziet}
\title{Create a hugo diziet site}
\usage{
create_hugodown_diziet(
  path = ".",
  open = rlang::is_interactive(),
  rstudio = rstudioapi::isAvailable()
)
}
\arguments{
\item{path}{Path to create site}

\item{open}{Open new site after creation?}

\item{rstudio}{Create RStudio project?}
}
\description{
Create a hugo diziet site
}
