% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_hugodown.R
\name{create_hugodown_slum}
\alias{create_hugodown_slum}
\title{Create a hugodown slum site}
\usage{
create_hugodown_slum(
  path = ".",
  open = rlang::is_interactive(),
  rstudio = rstudioapi::isAvailable(),
  publish_dir = "public"
)
}
\arguments{
\item{path}{Path to create site}

\item{open}{Open new site after creation?}

\item{rstudio}{Create RStudio project?}

\item{publish_dir}{Folder to publish to (typically "public" or "docs")}
}
\description{
Create site with hugodown
}
