% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation.R
\name{new_site}
\alias{new_site}
\title{Creates new blogdown site with slum theme}
\usage{
new_site(dir = ".", ...)
}
\arguments{
\item{dir}{Where to create the new blog}

\item{...}{Arguments to be passed to blogdown::new_site}
}
\description{
Creates new blogdown site with slum theme
}
\details{
This is just a wrapper to blogdown::new_site
}
