% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation.R
\name{slum_new}
\alias{slum_new}
\title{Creates new blogdown site with slum theme}
\usage{
slum_new(dir = ".", theme = "djnavarro/hugo_slum",
  hostname = "github.com", ...)
}
\arguments{
\item{dir}{Where to create the new blog}

\item{theme}{Location of hugo theme}

\item{hostname}{Where to look for repo}

\item{...}{Other arguments to be passed to blogdown::new_site}
}
\description{
Creates new blogdown site with slum theme
}
\details{
This is just a wrapper to blogdown::new_site
}
