% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{slum_palette}
\alias{slum_palette}
\title{Retrieve palette information}
\usage{
slum_palette(palette)
}
\arguments{
\item{palette}{Character string with the name of the palette (e.g., "dark")}
}
\value{
A named character vector
}
\description{
Retrieve palette information
}
\details{
Searches the two standard locations (user static/css and theme
static/css) to find a slum palette with the appropriate name.
}
