% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{slum_palette_create}
\alias{slum_palette_create}
\title{Retrieve palette information}
\usage{
slum_palette_create(name, pagecolour, maintext, faded, highlight, lowlight,
  overwrite = FALSE)
}
\arguments{
\item{name}{Character string with the name of the new palette (e.g., "cerulean")}

\item{pagecolour}{Character string with hex code (e.g., "#333333") for the page colour}

\item{maintext}{Character string specifying the main text colour}

\item{faded}{Character string specifying the "faded" colour}

\item{highlight}{Character string for highlight (mostly in links)}

\item{lowlight}{Character string for lowlight (mostly in code)}

\item{overwrite}{Should an existing file be overwritten?}
}
\description{
Retrieve palette information
}
\details{
Writes a file to the user directory
}
