% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{slum_plot_background}
\alias{slum_plot_background}
\title{Specify background colour with knitr}
\usage{
slum_plot_background(palette = "dark", name = "pagecolour")
}
\arguments{
\item{palette}{The palette name (default = "dark")}

\item{name}{The colour (default = "pagecolour")}
}
\description{
Specify background colour with knitr
}
\details{
Convenience function that uses knitr to pass a background colour
argument to the graphics device
}
