% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation.R
\name{slum_tidy}
\alias{slum_tidy}
\title{Creates new blogdown site with slum theme}
\usage{
slum_tidy(dir = ".", here = TRUE, nojekyll = TRUE)
}
\arguments{
\item{dir}{Where to create the new blog}

\item{here}{Whether to create a .here file}

\item{nojekyll}{Whether to create a .nojekyll file}
}
\description{
Creates new blogdown site with slum theme
}
\details{
This is just a wrapper to blogdown::new_site
}
