% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_slum.R
\name{theme_slum}
\alias{theme_slum}
\title{Creates ggplot2 themes for the slum blogdown theme}
\usage{
theme_slum(palette = "dark", ...)
}
\arguments{
\item{palette}{A character string specifying a palette}

\item{...}{Additional arguments to be passed to theme_grey}
}
\value{
A ggplot2 theme
}
\description{
Creates ggplot2 themes for the slum blogdown theme
}
\details{
Makes a ggplot2 theme
}
