% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation.R
\name{new_slum}
\alias{new_slum}
\title{Creates new slumdown site}
\usage{
new_slum(dir, remote = FALSE, project = "rstudio", nojekyll = TRUE)
}
\arguments{
\item{dir}{Where to create the new blog}

\item{remote}{Install from github (remote = TRUE) or locally (remote = FALSE)}

\item{project}{String specifying whether to use an rstudio project
(project = "rstudio") or a .here file (project = "here")}

\item{nojekyll}{Add a nojekyll file?}
}
\description{
Creates new slumdown site
}
\details{
Create a new slumdown blog
}
