% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation.R
\name{new_slum}
\alias{new_slum}
\title{Creates new blogdown site with slum theme}
\usage{
new_slum(dir = ".", project = TRUE, nojekyll = TRUE, ...)
}
\arguments{
\item{dir}{Where to create the new blog}

\item{project}{Create a tidy slum}

\item{nojekyll}{Add a nojekyll file}

\item{...}{Other arguments to be passed to blogdown::new_site}
}
\description{
Creates new blogdown site with slum theme
}
\details{
This is just a wrapper to blogdown::new_site
}
