% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installation.R
\name{new_slum}
\alias{new_slum}
\title{Creates new slumdown site}
\usage{
new_slum(dir, remote = FALSE, project = TRUE, nojekyll = TRUE)
}
\arguments{
\item{dir}{Where to create the new blog}

\item{remote}{Install from github (remote = TRUE) or locally (remote = FALSE)}

\item{project}{Create files to specify location of project root?}

\item{nojekyll}{Add a nojekyll file?}
}
\description{
Creates new slumdown site
}
\details{
Create a new slumdown blog
}
